Added HeadPlanController

+++ src/main/java/com/mathpar/plans/services/HeadPlanService.java
@@ -3,10 +3,13 @@ package com.mathpar.plans.services;
 import com.mathpar.plans.entities.HeadPlan;
 import com.mathpar.plans.repositories.HeadPlanRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.Date;
+import java.util.Optional;
 
+@Service
 public class HeadPlanService {
 
     @Autowired
@@ -14,6 +17,11 @@ public class HeadPlanService {
 
     public HeadPlan getHeadPlanById(long headPlanId){
         return headPlanRepository.getById(headPlanId);
+    }
+
+    public HeadPlan findHeadPlanById(long headPlanId) {
+        Optional<HeadPlan> optional = headPlanRepository.findById(headPlanId);
+        return optional.orElse(null);
     }
 
     public HeadPlan createHeadPlan(String name, Date startDate, Date endDate){
@@ -26,6 +34,9 @@ public class HeadPlanService {
 
     @Transactional
     public void purgeHeadPlanById(long headPlanId){
-        headPlanRepository.deleteById(headPlanId);
+        HeadPlan headPlan = findHeadPlanById(headPlanId);
+        if (headPlan != null) {
+            headPlanRepository.deleteById(headPlanId);
+        }
     }
 }
